third_party/blink/web_tests/platform/linux-chrome/external/wpt/wai-aria/role/tab-roles-expected.txt

This is a testharness.js-based test.
[FAIL] orphan button with tab role
  assert_equals: <button role="tab" data-testname="orphan button with tab role" data-expectedrole="button" class="ex">x</button> expected "button" but got "tab"
Harness: the test ran to completion.
